chore: cyclic dependency between metrcis and tasks resolved
