+++ src/eval_framework/tasks/base.py
@@ -19,10 +19,11 @@ from eval_framework.metrics.efficiency.bytes_per_sequence_position import (
     SequencePositionsLoglikelihood,
 )
 from eval_framework.metrics.efficiency.token_counters import TokenCounts
+from eval_framework.shared.errors import raise_errors
 from eval_framework.shared.types import BaseMetricContext, Completion, Error, RawCompletion
 from eval_framework.tasks.dataset_revisions import pinned_revision
 from eval_framework.tasks.markdown_doc import markdown_doc as render_markdown_doc
-from eval_framework.tasks.utils import classproperty, raise_errors
+from eval_framework.tasks.utils import classproperty
 from template_formatting.formatter import BaseFormatter, Message, Role
 
 if TYPE_CHECKING:
